requirements.txt
Flask==3.0.2
flask-cors==4.0.0
SpeechRecognition==3.10.1
python-dotenv==1.0.1
